feat(ui): update notification type and move mock object
